Use recommended validator set; add TestKit mock
Replace random selection in recommendedCurrentValidator with a deterministic lookup using recommendedValidators.validatorsSet. Add a new StakeSceneViewModel+TestKit.swift providing a mock factory for tests. Update StakeSceneViewModelTests to use the TestKit, adjust imports, remove the inline mock/TODO, and add a unit test for recommendedCurrentValidator to assert the correct validator is returned.
